generics: introduce As/AsFn and SliceAs/SliceAsFn
As() uses a standard type cast to convert a value to a different type.

SliceAs() uses a standard type cast and returns the subset of the values
that could be converted to the different type.

AsFn() is like As() but uses a helper function.

SliceAsFn() is like SliceAs() but uses a helper function, and it
is a simpler version of SliceMap()

Signed-off-by: Alejandro Mery <amery@geeks.cl>